用户页面修改 删除管理员bug，复选框bug

+++ web-ssm/src/main/java/com/arvato/oms/controller/UserController.java
@@ -132,7 +132,11 @@ public class UserController
         {
             list.add(Integer.valueOf(listString.get(i)));
         }
+        if(list.contains(6)){
+            return -1;
+        }else {
+            return userModelService.deleteUserByIds(list);
+        }
 
-        return userModelService.deleteUserByIds(list);
     }
 }
